requirements.txt
flask
mysqlclient
flask-sqlalchemy
flask-Mail
flask-MySQLdb
flask_redis
SQLAlchemy
opencv-python
pymysql
paho-mqtt
pillow
uwsgi
pika
pyyaml